update(Internals): Improve SOLID
Also removed renderer factories.
