+++ fastapi_forge/render/__init__.py
@@ -1,5 +1,3 @@
-from typing import Any
-
 from .engines.jinja2_engine import Jinja2Engine
 from .manager import RenderManager
 from .registry import RendererRegistry
@@ -10,5 +8,5 @@ def create_jinja_render_manager(project_name: str) -> RenderManager:
     jinja_engine.add_global("project_name", project_name)
     return RenderManager(
         engine=jinja_engine,
-        factories=RendererRegistry.get_factories(),
+        renderers=RendererRegistry.get_renderers(),
     )
