+++ fastapi_forge/render/manager.py
@@ -1,30 +1,27 @@
 from .engines import TemplateEngine
-from .factories import RendererFactory
-from .renderers.protocols import Renderer
+from .renderers import Renderer
 
 
 class RenderManager:
     def __init__(
         self,
         engine: TemplateEngine,
-        factories: dict[str, type[RendererFactory]],
+        renderers: dict[str, type[Renderer]],
     ):
         self.engine = engine
-        self.factories = factories
+        self.renderers = renderers
         self._renderers: dict[str, Renderer] = {}
 
     def get_renderer(self, renderer_type: str) -> Renderer:
         """Get a renderer instance for the specified type."""
-        if renderer_type not in self.factories:
+        if renderer_type not in self.renderers:
             raise ValueError(
-                f"No factory registered for renderer type: {renderer_type}"
+                f"No renderer registered for renderer type: {renderer_type}"
             )
 
         if renderer_type not in self._renderers:
-            factory_class = self.factories[renderer_type]
-            factory_instance = factory_class()
-            self._renderers[renderer_type] = factory_instance.create(
-                self.engine,
-            )
+            renderer_class = self.renderers[renderer_type]
+            renderer_instance = renderer_class(self.engine)
+            self._renderers[renderer_type] = renderer_instance
 
         return self._renderers[renderer_type]
