--- fastapi_forge/render/renderers/dao_renderer.py
@@ -1,15 +1,0 @@
-from typing import Any
-
-from fastapi_forge.schemas import Model
-
-from ..engines.protocols import TemplateEngine
-from ..templates import DAO_TEMPLATE
-from .protocols import Renderer
-
-
-class DAORenderer(Renderer):
-    def __init__(self, engine: TemplateEngine):
-        self.engine = engine
-
-    def render(self, model: Model, **kwargs: Any) -> str:
-        return self.engine.render(DAO_TEMPLATE, {"model": model, **kwargs})
